Use headerless embed query parameter for the oEmbed iframe
