+++ apps/badgrsocialauth/utils.py
@@ -1,25 +1,9 @@
 import urllib
-import urlparse
 
 from django.http import HttpResponseRedirect
-
-from rest_framework.response import Response
-from rest_framework.status import HTTP_500_INTERNAL_SERVER_ERROR
 from rest_framework.authentication import TokenAuthentication
 
 from mainsite.models import BadgrApp
-
-
-def set_url_query_params(url, **kwargs):
-    """
-    Given a url, possibly including query parameters, return a url with the given query parameters set, replaced on a
-    per-key basis.
-    """
-    url_parts = list(urlparse.urlparse(url))
-    query = dict(urlparse.parse_qsl(url_parts[4]))
-    query.update(kwargs)
-    url_parts[4] = urllib.urlencode(query)
-    return urlparse.urlunparse(url_parts)
 
 
 def get_session_verification_email(request):
