+++ apps/badgrsocialauth/views.py
@@ -9,9 +9,10 @@ from django.http import HttpResponseBadRequest, HttpResponseForbidden
 from django.views.generic import RedirectView
 from rest_framework.exceptions import AuthenticationFailed
 
-from badgrsocialauth.utils import set_url_query_params, set_session_badgr_app, get_session_badgr_app, \
-    get_session_verification_email, set_session_authcode
+from badgrsocialauth.utils import (set_session_badgr_app, get_session_badgr_app,
+                                   get_session_verification_email, set_session_authcode,)
 from mainsite.models import BadgrApp
+from mainsite.utils import set_url_query_params
 
 
 class BadgrSocialLogin(RedirectView):
