+++ apps/mainsite/utils.py
@@ -7,6 +7,7 @@ import StringIO
 import base64
 import hashlib
 import re
+import urllib
 import urlparse
 import uuid
 
@@ -157,3 +158,15 @@ def list_of(value):
     elif isinstance(value, list):
         return value
     return [value]
+
+
+def set_url_query_params(url, **kwargs):
+    """
+    Given a url, possibly including query parameters, return a url with the given query parameters set, replaced on a
+    per-key basis.
+    """
+    url_parts = list(urlparse.urlparse(url))
+    query = dict(urlparse.parse_qsl(url_parts[4]))
+    query.update(kwargs)
+    url_parts[4] = urllib.urlencode(query)
+    return urlparse.urlunparse(url_parts)
